requirement.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-community
langchain-text-splitters
requests
defusedxml
bs4
pymupdf
arxiv
langchain-huggingface
langchain-pinecone
langchain_weaviate
langchain_google_genai
wikipedia
youtube_search
langgraph
langchain-core
typing-extensions
chromadb
duckduckgo_search
yfinance
langchain-openai
